refactor: extract giftFlowerService
