+++ src/main/java/com/twoCube/gifts/service/GiftService.java
@@ -7,8 +7,6 @@ import org.springframework.web.multipart.MultipartFile;
 import java.util.List;
 
 public interface GiftService {
-    FlowerResponse getRandomFlower(Member member);
-    Long createFlower(FlowerRequest flowerRequest, Member member);
     List<PillListResponse> getPillList();
     Long createPill(PillRequest pillRequest, Member member);
     Long createPolaroid(MultipartFile polaroid, String content, Member member);
