+++ src/main/java/com/twoCube/gifts/service/GiftServiceImpl.java
@@ -17,28 +17,12 @@ import java.util.List;
 @RequiredArgsConstructor
 public class GiftServiceImpl implements GiftService{
 
-    private final FlowerRepository flowerRepository;
     private final GiftFlowerRepository giftFlowerRepository;
     private final PillRepository pillRepository;
     private final GiftPillRepository giftPillRepository;
     private final S3Uploader s3Uploader;
     private final GiftPolaroidRepository giftPolaroidRepository;
 
-
-    @Override
-    public FlowerResponse getRandomFlower(Member member) {
-        Flower flower = flowerRepository.findFlowerByRand();
-        return new FlowerResponse(flower);
-    }
-
-    @Override
-    public Long createFlower(FlowerRequest flowerRequest, Member member) {
-        Couple couple = member.getCouple();
-        Flower flower = flowerRepository.getById(flowerRequest.getFlowerId());
-        GiftFlower giftFlower = flowerRequest.toEntity(member, couple, flower);
-        giftFlowerRepository.save(giftFlower);
-        return giftFlower.getId();
-    }
 
     @Override
     public List<PillListResponse> getPillList() {
